requirements.txt
anthropic>=0.40.0
picamera2>=0.3.12
numpy
Pillow
python-dotenv
openai>=1.0.0
sounddevice>=0.4.6
